Throw lock exceptions

+++ app/Models/Stateless/LibraryNode.php
@@ -4,11 +4,12 @@
 namespace App\Models\Stateless;
 
 
+use App\Exceptions\CouldNotAcquireLockException;
 use App\Models\Document;
 use App\Models\Library;
 use Carbon\Carbon;
-use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Str;
+use Symfony\Component\Mime\MimeTypes;
 
 class LibraryNode implements \JsonSerializable
 {
@@ -86,6 +87,11 @@ class LibraryNode implements \JsonSerializable
         return $flags;
     }
 
+    public function getMime(): ?string
+    {
+        return MimeTypes::getDefault()->guessMimeType($this->getAbsolutePath());
+    }
+
     public function jsonSerialize()
     {
         return [
@@ -101,6 +107,9 @@ class LibraryNode implements \JsonSerializable
         ];
     }
 
+    /**
+     * @throws CouldNotAcquireLockException
+     */
     public function moveToTrash()
     {
         $trashPath = $this->library->getAvailableTrashPath($this->getFileInfo()->getBasename());
@@ -114,6 +123,10 @@ class LibraryNode implements \JsonSerializable
 
             /** @var Document $document */
             foreach ($oldDocuments as $document) {
+                if (!$document->makeLock(3)->get()) {
+                    throw new CouldNotAcquireLockException();
+                }
+
                 $document->path = Str::of($document->path)->replaceFirst($this->path, $trashPath);
                 $document->trashed_at = Carbon::now();
                 $document->save();
@@ -121,6 +134,10 @@ class LibraryNode implements \JsonSerializable
         } else {
             $document = $this->getDocument();
             if ($document) {
+                if (!$document->makeLock(3)->get()) {
+                    throw new CouldNotAcquireLockException();
+                }
+
                 $document->path = $trashPath;
                 $document->trashed_at = Carbon::now();
                 $document->save();
@@ -132,9 +149,14 @@ class LibraryNode implements \JsonSerializable
 
     public function delete()
     {
-        unlink($this->fileInfo->getRealPath());
         if ($document = $this->getDocument()) {
+            if (!$document->makeLock(3)->get()) {
+                throw new CouldNotAcquireLockException();
+            }
+
             $document->delete();
         }
+
+        unlink($this->fileInfo->getRealPath());
     }
 }
